requirements.txt
# requirements.txt

Django==5.1
djangorestframework==3.15.2
PyJWT==2.9.0
python-dotenv==1.0.1
django-filter==24.3
